Move the generic full stack code tests into separate folder
(Motivation: a new horse64-based testing tool that will be made
later shall be able to pick them up, and handle them in CI for
all the cross compiled platforms. Running them through
test_vmexec.c will remain as a first line of defense though,
for testing without needing a working horse64 to run tests in
the first place, to better diagnose super broken builds.)

DCO-1.1-Signed-off-by: Ellie <kittens@wobble.ninja>

+++ horse64/nonlocale.h
@@ -58,7 +58,7 @@ ATTR_UNUSED static inline double h64atof(const char *s) {
     #endif
 }
 
-ATTR_UNUSED static inline long long int h64atoll(const char *s) {
+ATTR_UNUSED static inline int64_t h64atoll(const char *s) {
     return h64strtoll(s, NULL, 10);
 }
